+++ Assets/Scripts/Enemies/Enemy_zombie_Kenny/Enemy_zombie_Kenny_damage.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Enemy_zombie_Kenny_damage : Damage {
 
-	public override void DefaultDamage (float damage, float direction)
+	public override void DefaultDamage (float damage, int direction)
 	{
 		ReduceHP (damage);
 		conditions.EnableStun (direction);
