+++ Assets/Scripts/Enemies/Enemy_zombie_Rubick/Enemy_zombie_Rubick_damage.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Enemy_zombie_Rubick_damage : Damage {
 	
-	public override void DefaultDamage (float damage, float stunDirection) {
+	public override void DefaultDamage (float damage, int stunDirection) {
 		conditions.EnableStun (stunDirection);
 		ReduceHP (damage);
 		anim.SetTrigger ("attackable");
