+++ Assets/Scripts/Enemies/Enemy_zombie_Zheron/Enemy_zombieZheron_damage.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Enemy_zombieZheron_damage : Damage {
 	
-	public override void DefaultDamage(float damage, float direction) {
+	public override void DefaultDamage(float damage, int direction) {
 	
 		bool backToTheEnemy = BackToTheEnemyCheck (direction);
 		//Если включен блок
