+++ Assets/Scripts/Enemies/Enemy_zombie_invoked_zombie/Enemy_zombie_invokedZombie_damage.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Enemy_zombie_invokedZombie_damage :  Damage {
 
-	public override void DefaultDamage (float damage, float stunDirection) {
+	public override void DefaultDamage (float damage, int stunDirection) {
 		conditions.EnableStun (stunDirection);	
 		unit.health = 0f;
 		conditions.UnitDie ();
